src/backend/requirements.txt
fastapi==0.129.0
pydantic==2.12.5
python-dotenv==1.2.1
SQLAlchemy==2.0.46
uvicorn==0.40.0
psycopg2-binary==2.9.11
email-validator==2.2.0